+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Documents changes that result in:
 
 ## Unreleased
 
+- get_contracts_deployment_info() returns None instead of raising a ValueError when no deployment file is found.
 - Deploy 0.4.0 version on Goerli
 - [#853](https://github.com/raiden-network/raiden-contracts/pull/853) add chain_id in the IOU claims